Implement BookingList and FarmerDetails components; integrate ViewFarmerContext for managing active farmer state and enhance routing in Bookings component.
